requirements.txt
python-docx==0.8.11
PyQt5==5.15.9
PyInstaller==5.13.0
